+++ src/store/constants/apiResponse.constants.js
@@ -1,7 +1,3 @@
 export const OPEN_RESPONSE_MODAL = 'OPEN_RESPONSE_MODAL'
 export const CLOSE_RESPONSE_MODAL = 'CLOSE_RESPONSE_MODAL'
-export const OPEN_CONFIRMATION_MODAL = 'OPEN_CONFIRMATION_MODAL'
-export const CLOSE_CONFIRMATION_MODAL = 'CLOSE_CONFIRMATION_MODAL'
 export const CLEAR_CONFIRMATION_MODAL = 'CLEAR_CONFIRMATION_MODAL'
-export const OPEN_TOASTER = 'OPEN_TOASTER'
-export const CLOSE_TOASTER = 'CLOSE_TOASTER'
