+++ src/store/reducers/apiResponse.reducer.js
@@ -1,14 +1,8 @@
 import {
   OPEN_RESPONSE_MODAL,
   CLOSE_RESPONSE_MODAL,
-  OPEN_CONFIRMATION_MODAL,
-  CLOSE_CONFIRMATION_MODAL,
-  CLEAR_CONFIRMATION_MODAL,
-  OPEN_TOASTER,
-  CLOSE_TOASTER
+  CLEAR_CONFIRMATION_MODAL
 } from '../constants'
-
-import setToasterMessage from '../../helpers/setToasterMessage'
 
 export const confirmationModalState = {
   actionType: '',
@@ -51,51 +45,10 @@ export default function reducer (state = initialState, action) {
         isOpen: true
       }
     }
-    case OPEN_CONFIRMATION_MODAL: {
-      const { actionType, api, apiName, message, isRedirect, status, title } = action.payload
-
-      return {
-        ...state,
-        confirmationModal: {
-          actionType,
-          api,
-          apiName,
-          message,
-          isRedirect,
-          status,
-          title,
-          needConfirm: true
-        }
-      }
-    }
-    case OPEN_TOASTER: {
-      return {
-        ...state,
-        toaster: {
-          isOpen: true,
-          message: setToasterMessage(state.confirmationModal)
-        }
-      }
-    }
-    case CLOSE_CONFIRMATION_MODAL: {
-      return {
-        ...state,
-        confirmationModal: {
-          ...state.confirmationModal,
-          needConfirm: false
-        }
-      }
-    }
     case CLEAR_CONFIRMATION_MODAL: {
       return {
         ...state,
         confirmationModal: confirmationModalState
-      }
-    }
-    case CLOSE_TOASTER: {
-      return {
-        ...state,
-        toaster: toasterState
       }
     }
     case CLOSE_RESPONSE_MODAL: {
